Further refactoring of Identiier to match spec...
I didn't notice that Identifier explicitly outlaws
period (.) characters in most cases, so have brought back
the idea of having two types of identifiers, but done it
slightly differently this time.

We now have Identifier & ComplexReferenceIdentifier.

The latter is used only by LookupExpression and subclasses.
This simplifies some of the existing variable dereferencing
logic, so is good.