Minor logic tweak to CandidateSessionStarter...
We now check to see whether the candidate already has
a non-terminated session on the given Deliverable. If there
is one, we'll reconnect to it (or the most recent one). Otherwise
we'll create a new session. This should help with situations
where the candidate clicks on a start button twice, or something
like that.