Helpers for M1->M2 & post-pilot data migration.
This is only of interest to David. I've committed this here
for posterity and will remove immediately afterwards.

The migration from M1->M2 involved a big DB schema change, as well
as a merge of the test data Sue Milne had deployed to dev for the
test piloting in December 2012. That makes it all a bit complicated!