WIP: multiple testParts & preConditions
This is a first sketch of support for multiple testParts,
and the evaluation of preConditions on them. It will need further
refinement. I've also added a new QtiCandidateStateException
for reporting state precondition failures and such-like; this
will be worked into the existing code as I complete the changes
to test navigation.