Further tweak to HEAD~2
I now realise the primary key generation issue when
deploying to dev was down to the sequences in the DB
being out of sync with the @SequenceGenerator annotations.

These need to be manually synced, which I have included here.
I expect to make some further tweaks for the next snapshot.