Fixed TSController: advancing items in L/S mode
This was previously ending the item session, which was the
wrong behaviour in S mode. The logic now behaves better
(suspending the session instead of ending it).

NB: I changed the name of this method to advanceItemLinear()
as endItemLinear() is no longer accurate. The corresponding
mayEndItemLinear() has been renamed as mayAdvanceItemLinear()

I have added a unit test for this.