URL of my hosted QTIWorks installs updated.
The final 1.0 release will use webapps.ph.ed.ac.uk/... instead of
www2.ph.ed.ac.uk/... in future. The old address will continue to be
supported so that existing LTI links don't break; though visits to
the static parts of the site will redirect to the new domain if
not there already.