Change to LTI result return logic
This is in light of recent changes that allow multiple LTI
results to be returned for a single CandidateSession. We now
check each batch for multiple results in discard all but the last
one. Provided the result returning is always called serially (which it
currently is due to the way it is invoked via ScheduledService), then
this will ensure that the final result always gets back, or at least
tries more than any other result to get back.