Fixed issue #35: numeric OV initial values
The info model said that *numeric* OVs should be set to 0 if no default
is provided. I had been setting these to NULL. This commit fixes this.

Note that this fix applies to OVs in both items and tests, though the
info model not completely clear on this as it's presented in the context
of item variables and item sessions at this point. I think test ought
to behave in this way too.