Delivery deletion now removes LTI link candidates
Recall that LTI link candidates are created especially for a delivery.
We now delete all of these users when a delivery is deleted, which is
the correct thing to do with regards to privacy and data protection.
(The original behavious was to keep the candidates in the DB, which
was an oversight.)