Improvement to candidate session launch workflow
I have moved the candidate session entry to a separate step after the
session is first instantiated. This allows an additional page to be
presented to the candidate which uses some JS to invoke the
entry/re-entry into the session. Candidates using the back button
will end up back at this page, which automatically pushes them back
into the session again.

Needs a bit of tidying, which will be in the next few commits