Some minor test rendering tweaks
Moved the test session control further up so that it is nearer the
item control buttons, hopefully making it easier for candidates to
notice them. Also did a bit of tidying to the confirm text when passing
through a point of no return in the test (e.g. end test part). Plan to
make these use nicer dialogs next, though.