Finished initial candidate refactor. Needs tested
The new candidate session management has been refactored and joined up
with the existing MVC and service layers.

I have tested with the demo & instruction functionality and this is
working correctly.

Still need to test with LTI, which I can't do until I'm back in the
office. (At home presently.)