Fixed linear nav case not covered in commit b2c4991
The commit b2c4991 only improved the timing of session finish in the
case of non-linear navigation. This commit does the same for linear
navigation, by finishing the session when the final item in the final
testPart is ended.

I have also added a new FINISH_FINAL_ITEM constant to
CandidateTestEventType to make it easy to see when the final item in a
testPart has been ended.