Minor refactoring to AssessmentManagerService Exceptions
The CannotChangeAssessmentTypeException and
IncompatiableDeliverySettingsException classes have been replaced with a
multi-purpose IllegalManagementOperationException, which behaves
similarly to PrivilegeException and
AssessmentPackageDataImportException.

Additionally, some logic has been added to ensure the manual operations
on Deliveries are only permitted on USER_CREATED deliveries. The new
excpetion class is used to report any violations of this.