New deletion logic for CandidateSessions & Assessments
I have rewritten the logic in DataDeletionService for deleting
CandidateSessions and Assessments from the DB. This no longer relies on
entity cascades and instead performs a number of bulk deletions,
performed in a suitable order to avoid constaint violations.

This makes bulk deletions *much* faster than before. However there are a
now a few more deletion queries in the various Entity objects, which is
a bit noisy.

I have also added some new actions to the Engine Manager to allow
certain deletions to be invoked that way too.