Added manager action for doing a DB schema update
Following the upgrade of Hibernate and plans to add some further indexes
to the schema, I've created an explicit manager action for performing a
database schema update.

This should be invoked when doing a version upgrade.

HOWEVER! Note that Hibernate 5 introduces a new naming convention of
indexes and foreign key constraints. So the schema upgrade going to the
next new release needs will introduce duplicate constraints & indexes,
so we'll need to get users to remove "old" keys.