Added mandatory Privacy Policy link
As preparation for the GDPR, I've added a configurable (but mandatory)
link to an external Privacy Policy page for your QTIWorks instance.

The link is specified via qtiworks-deployment.properties.

The Edinburgh instance of QTIWorks will link to an updated version of
the Privacy, Security & Cookies page in the QTIWorks wiki on GitHub.