Issue #77: Hopeful fix + regression test
The CorrectResponse class was assuming that it would only ever be a
child of a ResponseDeclaration, whereas it can also be a child of
responseVariable. This was leading to a ClassCastException when loading
an <assessmentResult> XML that contained a <correctResponse>.

This commit refactors the class so that it works in both contexts, and
I've added a regression test to ensure that we can safely load an
<assessmentResult> containing a <correctResponse>.

As I've not looked at this code for a long time, I can't guarantee this
change won't have any unexpected side effects in certain cases.
So further testing would be beneficial before merging back into the
master branch.